Handle missing namespace packages (#28)
* Handle missing namespace packages

Closes #27

+++ src/whl.py
@@ -40,6 +40,9 @@ def configure_reproducible_wheels():
 
 def _create_nspkg_init(dirpath):
     """Creates an init file to enable namespacing"""
+    if not os.path.exists(dirpath):
+        # Handle missing namespace packages by ignoring them
+        return
     nspkg_init = os.path.join(dirpath, "__init__.py")
     with open(nspkg_init, "w") as nspkg:
         nspkg.write("__path__ = __import__('pkgutil').extend_path(__path__, __name__)")
